requirements.txt
pypresence
pywebview